changed 'Number' types to 'Integer' for Analog IO

+++ pi4j-api/src/main/java/com/pi4j/annotation/ShutdownValue.java
@@ -35,5 +35,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.FIELD})
 public @interface ShutdownValue {
-    double value();
+    int value();
 }
